bot refactoring
