+++ bot/.env.template
@@ -1,0 +1,9 @@
+TOKEN_TELEGRAM=
+TOKEN_NOTION=
+
+RABBITMQ__DEFAULT_HOST=localhost
+RABBITMQ__DEFAULT_USER=guest
+RABBITMQ__DEFAULT_PASS=guest
+RABBITMQ__WEB_PORT=15672
+RABBITMQ__QUEUE_PORT=5672
+RABBITMQ__TELEGRAM_QUEUE=telegram
