+++ bot/Makefile
@@ -1,0 +1,10 @@
+flake8:
+	poetry run flake8 src main.py
+
+isort:
+	poetry run isort src main.py
+
+black:
+	poetry run black src main.py
+
+check: isort black src main.py
